fixed cases to work both with and without Log::Dispatch

+++ t/018Init.t
@@ -8,6 +8,14 @@ use strict;
 use Log::Log4perl;
 use File::Spec;
 
+my $WORK_DIR = "tmp";
+if(-d "t") {
+    $WORK_DIR = File::Spec->catfile(qw(t tmp));
+}
+unless (-e "$WORK_DIR"){
+    mkdir("$WORK_DIR", 0755) || die "can't create $WORK_DIR ($!)";
+}
+
 my $testfilea = File::Spec->catfile(qw(t tmp test18a.log));
 unlink $testfilea if (-e $testfilea);
