+++ t/027Watch2.t
@@ -11,6 +11,14 @@ use Log::Log4perl;
 use Log::Log4perl::Appender::TestBuffer;
 use File::Spec;
 
+my $WORK_DIR = "tmp";
+if(-d "t") {
+    $WORK_DIR = File::Spec->catfile(qw(t tmp));
+}
+unless (-e "$WORK_DIR"){
+    mkdir("$WORK_DIR", 0755) || die "can't create $WORK_DIR ($!)";
+}
+
 my $testconf= File::Spec->catfile(qw(t tmp test27.conf));
 unlink $testconf if (-e $testconf);
